Move ClusterResources to config-provision
